feat: run app without cuda core NVIDIA
This small adjustment makes it possible to run the application without
needing to use NVIDIA CUDA cores.

This allows more machines that don't have dedicated graphics cards or
have incompatible GPUs to use the software.

The adjustment simply involves passing the "device=cpu" parameter in
the constructor, which is already supported by Kokoro.